+++ .runtime/hooks/post_load.sh
@@ -1,0 +1,2 @@
+#/bin/bash
+/opt/ENV/bin/python manage.py load_lingo_fixtures --no-index