+++ arches_lingo/src/arches_lingo/api.ts
@@ -754,20 +754,19 @@ export const exportThesaurus = async (
     fileName?: string,
 ) => {
     const formData = new FormData();
-    formData.append("action", "start");
-    // moduleId normally fetched in context of BDM, but needed to get throough ETLManagerView POST
-    const lingoImporterModuleId = "4302e334-33ed-4e85-99f2-fdac7c7c32fa";
-    formData.append("module", lingoImporterModuleId);
     formData.append("resourceid", resourceid);
     formData.append("format", format);
     if (fileName) {
         formData.append("filename", fileName);
     }
-    const response = await fetch(arches.urls.etl_manager, {
-        method: "POST",
-        headers: { "X-CSRFTOKEN": getToken() },
-        body: formData,
-    });
+    const response = await fetch(
+        generateArchesURL("arches_lingo:api-lingo-export"),
+        {
+            method: "POST",
+            headers: { "X-CSRFTOKEN": getToken() },
+            body: formData,
+        },
+    );
     try {
         const parsed = await response.json();
         if (response.ok) {