+++ docker/hooks/post_load.sh
@@ -1,2 +1,0 @@
-#/bin/bash
-/opt/ENV/bin/python manage.py loaddata /opt/arches-lingo/tests/fixtures/data/FISH_Example_Thesauri.json.xz